Fix minor Budget

+++ src/components/bill/BillPatti.jsx
@@ -9,6 +9,12 @@ const BillPatti = ({ amount, date, id, fetchPaymentArray }) => {
 
   //Delete Bill Patti
   const deleteBillPatti = async () => {
+
+    const confirmDelete = window.confirm(
+        "Are you sure you want to delete this transaction?"
+    );
+    if(!confirmDelete) return;
+
     setLoading(true);
     try {
       const res = await axios.delete(`${server}/payment/${id}`, {
